# Add JsonStringArrayExtractScalar function