--- a/core/trino-main/src/main/java/io/trino/operator/scalar/ArrayTransformFunction.java
+++ b/core/trino-main/src/main/java/io/trino/operator/scalar/ArrayTransformFunction.java
@@ -84,10 +84,11 @@ public final class ArrayTransformFunction
     }
 
     public static final ArrayTransformFunction ARRAY_TRANSFORM_FUNCTION = new ArrayTransformFunction();
+    public static final String ARRAY_TRANSFORM_NAME = "transform";
 
     private ArrayTransformFunction()
     {
-        super(FunctionMetadata.scalarBuilder("transform")
+        super(FunctionMetadata.scalarBuilder(ARRAY_TRANSFORM_NAME)
                 .signature(Signature.builder()
                         .typeVariable("T")
                         .typeVariable("U")