… core/trino-main/src/test/java/io/trino/sql/query/QueryAssertions.java