<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

Optimize common case of `transform(cast(json_parse), json_extract_scalar)` to be single and more performant call of `JsonStringArrayExtractScalar`.

Example query:
```SQL
WITH t AS (SELECT a.id as id, a.data AS data FROM table_a AS a) 
SELECT b.data FROM table_b AS b, t AS a 
WHERE b.id = a.id AND  
CASE WHEN b.data <> 'second' THEN contains(transform(cast(json_parse(a.data) as array<json>), (x) -> json_extract_scalar("x", '$.name')), b.data) 
ELSE false END;
```


With change:
```
     │   CPU: 8.58s (94.13%), Scheduled: 8.93s (69.96%), Blocked: 1.89s (94.59%), Output: 3492632 rows (34.65MB)                                        
```

Without:
```
     │   CPU: 16.49s (95.06%), Scheduled: 17.24s (82.05%), Blocked: 2.81s (92.72%), Output: 3492632 rows (34.65MB)            
```
<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or is docs only, and no release notes are required.
( ) Release notes are required. Please propose a release note for me.
(X) Release notes are required, with the following suggested text:

```markdown
# General
* Improve performance of parsing json array and extracting scalar
```